First step towards checking result of build job
Added checking of job status to function build_easystack_from_pr.
Whenever a job has finished the function process_job_result is called.
This then looks into artifacts produced by the job and adds a comment to
the PR (the actual adding is not yet happening).

Both implementations are just a first attempt to get it working. Likely
there is lots of potential for optimisation.