Adjusted method to call create_tarball.sh
If EESSI pilot is only available via a container, some settings may not
be derived outside. Hence, we need to change the way the
create_tarball.sh script is called and determine these settings inside
the script (assuming the EESSI pilot repo or a clone is available).

This requires a change in the software layer as well.