Enabling building for generic targets
Added one parameter to the script. All except the first parameter are
added as parameter to the calls for installing the software layer and
for creating a tarball. For generic targets, this script would be called
with the additional parameter --generic, which is then passed through to
`install_software_layer.sh` and `create_tarball.sh` as additional
parameter "$@".