Added handling of new config parameters
1) cvmfs_customizations allows to customize the configuration of
CVMFS in a container. For example, CVMFS_HTTP_PROXY can be changed
for nodes which have access to the internet via a proxy only.

2) http_proxy and https_proxy are used if build nodes do not have
direct to the internet. Typically these are set in a job's
environment. However, entering the Gentoo Prefix (startprefix)
resets most of the environment variables. Hence, they have to be
set again. Here they are added as parameters to the Slurm build
script which passes them through to the install script of the
software layer. The install script needs to be modified to parse
these parameters because the Prefix is entered at a late stage
only.