Fixing non-bot jobs leaking into job manager loops
While non-bot jobs are not processed in process_new_job they could still leak into the list of known jobs via
```python
known_jobs = current_jobs

at the end of the main loop in the job manager. To fix this we collect a list of non-bot jobs and remove them from the dictionary current_jobs just before that is assigned to known_jobs. The function process_new_job was slightly modified to return False if it processes a non-bot job and True otherwise.